Change name plugin and remove old version
The current name generates bureaucratic problems with wordpress.org,
so it cannot be used to upload the plugin to their platform. To make the
change, it is necessary to deactivate and eliminate the previous
versions of the plugin, in this way the commerce will be able to
automatically see and update future updates.